+++ examples/simple/search_indexes/documents/__init__.py
@@ -1,3 +1,9 @@
 from .author import *
 from .book import *
 from .publisher import *
+
+__all___ = (
+    'AuthorDocument',
+    'BookDocument',
+    'PublisherDocument',
+)